pyproject.toml
[project]
name = "fashionclip"
version = "0.1.0"
description = "Fashion retrieval and outfit recommendation with CLIP + LLM reranking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-multipart>=0.0.9",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pillow>=10.4.0",
  "torch>=2.3.0",
  "transformers>=4.44.0",
  "scikit-learn>=1.5.0",
  "tqdm>=4.66.0",
  "openai>=1.40.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
